Move headers command into own file
